requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.9.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
python-magic==0.4.27
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.55.0
shellingham==1.5.4
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0

# Logging
loguru>=0.7.3

# Authentication
authlib==1.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
bcrypt==5.0.0
aiosqlite==0.20.0
